builds_complete/v0.7.15_beta-candidate-0_COMPLETE/requirements.txt
python-chess==1.999
pygame==2.5.2
stockfish==3.28.0
